WIP: mac os app done completely in pony
After the generic objc unit tests run successfully, if this is not
being run on the CI server, then run the program as if it was a Mac
OS application (ie run NSApplicationMain()).  To do this we also
needed to add support for linking in a Info.plist file, which I added
to ponyc at the same time.